Fix original logo FOUC (flash of unstyled content)
Here is the problem. If javascript is available we hide original logo
and transition to it while user scrolls down the page.
While essential javascript asset and stylesheet are loading the logo can
be already rendered and apparently there occurs a FOUC (flash of unstyled
content) effect.

To fix this I'm using `gulp-inline-source` task that inlines essential
javascript and stylesheet file into the HTML markup right inside the <head> tag.
Fixed