Fix bug related to small window resize handler
When you first load page in window with width less than 480px
you won't get SVG stars effect (due to my poor knowdledge in how
to properly scale down svg groups without loosing their relational
position as a whole and not messing their future coordinates, so the
minimum width for stars is 480px as this is the fixed width of wireframe
that holds that stars in the end of transition) and as when resizing
the window the handler tries to update svg coordinates it throws
an error because svg objects were not initiated.

This fix adds a test for svg initialization in window resize handler
so if svg objects were not initiated window resize doesn't update
svg coordinates.